Add files via upload
In this task, you will create a PowerShell script. PowerShell enables administrators to perform administrative tasks on both local and remote systems. You will be expected to manage an Active Directory server and an SQL server within the PowerShell environment. This management will include the configuration and administration of the servers.